+++ src/components/Home/MapComponent/index.js
