+++ src/components/Home/NoEventsComponent/index.js
